+++ include/nc_runtime/pose3.h
@@ -15,7 +15,8 @@ public:
     Quaternion ori = Quaternion::identity();
 
     Pose3() = default;
-    Pose3(const Vector3& pos_, const Quaternion& ori_) : pos(pos_), ori(ori_) {}
+    forceinline Pose3(const Pose3& r) : pos(r.pos), ori(r.ori) {}
+    forceinline Pose3(const Vector3& pos_, const Quaternion& ori_) : pos(pos_), ori(ori_) {}
     static forceinline Pose3 makeWithPose2(const Pose2& pose2)
     {
         return Pose3(Vector3(pose2.pos.x, pose2.pos.y, 0), Quaternion::makeWithRPY(0, 0, pose2.ori));